python/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloud-sql-python-connector==1.18.5
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
multidict==6.7.0
mysql-connector-python==9.4.0
narwhals==2.8.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
yarl==1.22.0
